requirements.txt
folium==0.14.0
haversine==2.8.0
numpy==1.24.4
pandas==2.0.3
Pillow==9.5.0
Pillow==10.0.1
plotly==5.16.1
streamlit==1.26.0
streamlit_folium==0.14.0
